backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
websockets==12.0
pillow==10.2.0
python-multipart==0.0.6
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2
greenlet==3.0.3

